backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.0
jieba==0.42.1
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv==1.0.0
aiofiles==23.2.1

